Style changes in some pages
Style changes in home, login, sign in, password change pages with sidebar and navbar.

Also updated .tsx files of these pages
