+++ src/pages/StartPage.tsx
@@ -4,8 +4,8 @@ import bitescoutLogo11 from "../assets/bitescout-logo1-1.png";
 const StartPage = () => {
   return (
     <div className="start-page">
+      <img className="logo" src={bitescoutLogo11} alt="Logo" />
       <div className="content">
-        <img className="logo" src={bitescoutLogo11} alt="Logo" />
         <h1 className="catchphrase">Welcome to BiteScout</h1>
         <p className="subtitle">
           Your go-to platform for restaurant reservations
