+++ src/styles/Button.css
@@ -18,12 +18,10 @@
     align-items: center;
     justify-content: center;
     transition: background-color 0.3s ease, box-shadow 0.3s ease;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
 }
 
 .button__:hover {
     background-color: #176a4a;
-    box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);
 }
 
 .button__ p {
