+++ src/styles/LoginPage.css
@@ -1,33 +1,44 @@
 .login-page {
   width: 100vw;
   height: 100vh;
-  background: linear-gradient(to bottom right, #1d9265, #3ca674);
+  background: linear-gradient(
+    to bottom right,
+    #1d9265,
+    #3ca674
+  ); /* Daha uyumlu bir yeşil ton geçişi */
   display: flex;
-  justify-content: center;
-  align-items: center;
+  justify-content: flex-end;
+  align-items: center; /* Kart yukarı alınır */
   font-family: "Manrope", sans-serif;
-  color: white;
-  overflow: hidden;
+  padding-top: 80px; /* Header ile kart arasında daha az boşluk */
+  overflow:hidden;
 }
 
 .login-form-container {
-  width: 380px;
-  padding: 40px;
-  background-color: #ffffff;
-  border-radius: 20px;
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1); 
+  width: 600px; /* Daha geniş yapalım */
+  height: 32%;
+  padding: 30px;
+  background-color: #ffffff; /* Beyaz arka plan */
+  border-radius: 16px; /* Daha yuvarlak köşeler */
+  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2); /* Daha belirgin gölge */
   display: flex;
   flex-direction: column;
   align-items: center;
-  animation: fadeIn 0.7s ease-in-out;
-  backdrop-filter: blur(10px);
+  justify-content: center;
+  opacity: 0;
+  animation: fadeIn 0.5s ease-in-out; /* Yumuşak görünüm */
+  animation: loginSlide 0.5s ease-in-out forwards;
+  animation-delay: 1s;
+  margin-right: 200px;
 }
 
-.header__logo {
-  width: 140px;
-  height: auto;
-  margin-bottom: 30px;
-  align-self: center; 
+.login-page .header__logo {
+  width: 1200px !important;
+  height: auto !important;
+  margin-bottom: 20px;
+  margin-right: 210px;
+  align-self: center; /* Logoyu tam ortalar */
+  animation: logoFadeIn 1s ease-in-out;
 }
 
 .title {
@@ -41,32 +52,29 @@
   width: 100%;
   display: flex;
   flex-direction: column;
-  gap: 25px;
+  gap: 20px;
 }
 
 .input-group {
   display: flex;
   flex-direction: column;
-  text-align: left;
-  margin-bottom: 10px;
+  gap: 5px;
+  width: 100%;
+  padding-bottom: 20px;
 }
 
 .input-group label {
-  font-size: 16px;
-  font-weight: 600;
-  color: #555;
-  margin-bottom: 8px;
-  margin-left: 3px;
+  font-size: 18px;
+  margin-bottom: 10px;
+  font-weight: bold;
+  color: #cccccc;
 }
 
 .input-group input {
-  width: 94%;
-  padding: 12px;
+  padding: 30px;
   border: 1px solid #ddd;
-  border-radius: 12px;
-  font-size: 16px;
-  background-color: #f9f9f9;
-  transition: all 0.3s ease;
+  border-radius: 10px; /* Daha yuvarlak köşeler */
+  font-size: 24px;
 }
 
 .input-group input:focus {
@@ -78,7 +86,6 @@
 .forgot-password {
   text-align: right;
   width: 100%;
-  margin-bottom: 20px;
 }
 
 .forgot-password a {
@@ -92,20 +99,22 @@
 }
 
 .login-button {
-  padding: 14px;
-  font-size: 18px;
+  padding: 15px 100px 15px 100px;
+  margin-top: 25px;
+  width: 100%;
+  font-size: 24px;
   background-color: #1d9265;
-  color: white;
+  color: #fff;
   border: none;
-  border-radius: 12px;
-  font-weight: 700;
+  border-radius: 10px; /* Daha yuvarlak köşeler */
+  font-weight: bold;
   cursor: pointer;
-  transition: all 0.3s ease;
+  transition: background-color 0.3s ease;
 }
 
 .login-button:hover {
-  background-color: #146a48;
-  transform: scale(1.05);
+  background-color: #f6c825;
+  color: #1d9265;
 }
 
 .login-button:focus {
@@ -115,7 +124,7 @@
 
 .signup-link {
   margin-top: 20px;
-  font-size: 14px;
+  font-size: 18px;
   color: #333;
 }
 
@@ -137,5 +146,25 @@
   to {
     opacity: 1;
     transform: translateY(0);
+  }
+}
+
+@keyframes logoFadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes loginSlide {
+  from {
+    opacity: 0;
+    transform: translateX(150%);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0%);
   }
 }
