+++ src/styles/PasswordChangePage.css
@@ -1,103 +1,134 @@
 .password-change-page {
   width: 100vw;
   height: 100vh;
-  background: linear-gradient(to bottom right, #1d9265, #3ca674); 
+  background: linear-gradient(to bottom right, #1d9265, #3ca674); /* Yeşil tonlarında geçiş */
   display: flex;
-  justify-content: center;
-  align-items: center;
+  justify-content: flex-end;
+  align-items: center; /* Kart yukarı alınır */
   font-family: "Manrope", sans-serif;
+  padding-top: 80px; /* Header ile kart arasında daha az boşluk */
+  overflow:hidden;
   color: white;
-  overflow: hidden; 
 }
 
-.form-container {
-  width: 380px;
-  padding: 40px;
-  background-color: #ffffff; 
-  border-radius: 20px; 
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1); 
-  text-align: center;
-  animation: fadeIn 0.7s ease-in-out; 
-  backdrop-filter: blur(10px); 
-}
+  .form-container {
+    width: 600px; /* Daha geniş yapalım */
+    height: 22%;
+    padding: 0px 30px 30px 30px;
+    background-color: #ffffff; /* Beyaz arka plan */
+    border-radius: 16px; /* Daha yuvarlak köşeler */
+    box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2); /* Daha belirgin gölge */
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    opacity: 0;
+    animation: fadeIn 0.5s ease-in-out; /* Yumuşak görünüm */
+    animation: loginSlide 0.5s ease-in-out forwards;
+    animation-delay: 1s;
+    margin-right: 200px;
+  }
 
-.header__logo {
-  width: 140px;
-  height: auto;
-  margin-bottom: 30px;
-}
+  .password-change-page .header__logo {
+    width: 1200px;
+    height: auto;
+    margin-bottom: 20px;
+    margin-right: 210px;
+    align-self: center; /* Logoyu tam ortalar */
+    animation: logoFadeIn 1s ease-in-out; /* Logo için yumuşak animasyon */
+  }
 
-.title {
-  font-size: 28px;
-  font-weight: bold;
-  margin-bottom: 25px;
-  color: #333;
-}
+  .title {
+    font-size: 32px;
+    font-weight: bold;
+    margin-bottom: -5px;
+    color: #1d9265;
+  }
 
-.password-change-form {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 25px;
-}
+  .password-change-form {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+  }
 
-.input-group {
-  text-align: left;
-}
+  .input-group {
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
+    width: 100%;
+  }
 
-.input-group label {
-  font-size: 16px;
-  font-weight: 600;
-  color: #555;
-  margin-bottom: 8px;
-  margin-left: 3px;
-}
+  .input-group label {
+    font-size: 18px;
+    margin-bottom: 10px;
+    font-weight: bold;
+    color: #cccccc;
+  }
 
-.input-group input {
-  width: 94%;
-  padding: 12px;
-  border: 1px solid #ddd;
-  border-radius: 12px;
-  font-size: 16px;
-  background-color: #f9f9f9;
-  transition: all 0.3s ease;
-}
+  .input-group input {
+    padding: 30px;
+    border: 1px solid #ddd;
+    border-radius: 10px; /* Daha yuvarlak köşeler */
+    font-size: 24px;
+  }
 
-.input-group input:focus {
-  outline: none;
-  border-color: #1d9265;
-  background-color: #e6f9f0;
-}
+  .input-group input:focus {
+    outline: none;
+    border-color: #1d9265;
+  }
 
-.submit-button {
-  padding: 14px;
-  font-size: 18px;
-  background-color: #1d9265;
-  color: white;
-  border: none;
-  border-radius: 12px; 
-  font-weight: 700;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
+  .submit-button {
+    padding: 15px 100px 15px 100px;
+    margin-top: -20px;
+    width: 100%;
+    font-size: 24px;
+    background-color: #1d9265;
+    color: #fff;
+    border: none;
+    border-radius: 10px; /* Daha yuvarlak köşeler */
+    font-weight: bold;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+  }
 
-.submit-button:hover {
-  background-color: #146a48;
-  transform: scale(1.05); 
-}
+  .submit-button:hover {
+    background-color: #f6c825;
+    color: #1d9265
+  }
 
-.submit-button:focus {
-  outline: none;
-  box-shadow: 0 0 10px rgba(29, 146, 101, 0.5); 
-}
+  .submit-button:focus {
+    outline: none;
+    box-shadow: 0 0 10px rgba(29, 146, 101, 0.5); 
+  }
 
-@keyframes fadeIn {
-  from {
+  @keyframes fadeIn {
+    from {
       opacity: 0;
       transform: translateY(-20px);
-  }
-  to {
+    }
+    to {
       opacity: 1;
       transform: translateY(0);
+    }
   }
-}
+
+  @keyframes logoFadeIn {
+    from {
+      opacity: 0;
+    }
+    to {
+      opacity: 1;
+    }
+  }
+
+  @keyframes loginSlide {
+    from {
+      opacity: 0;
+      transform: translateX(150%);
+    }
+    to {
+      opacity: 1;
+      transform: translateX(0%);
+    }
+  }
